+++ src/settings/AdvancedSettingsXmlReader.cpp
@@ -174,6 +174,8 @@ void AdvancedSettingsXmlReader::loadGui()
     while (m_xml.readNextStartElement()) {
         if (m_xml.name() == "forceCache") {
             expectBool(m_settings.m_forceCache);
+        } else if (m_xml.name() == "stylesheet") {
+            m_settings.m_customStylesheet = m_xml.readElementText().trimmed();
         } else {
             skipUnsupportedTag();
         }